Move end reason and contract end wizard to contract_compassion module
