+++ contract_compassion/view/end_contract_wizard_view.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+    Copyright (C) 2014 Compassion (http://www.compassion.ch)
+    @author Emanuel Cino <ecino@compassion.ch>
+    The licence is in the file __openerp__.py
+-->
+<openerp>
+    <data>
+		<record id="end_contract_view" model="ir.ui.view">
+			<field name="name">end.contract.wizard.view</field>
+			<field name="model">end.contract.wizard</field>
+			<field name="arch" type="xml">
+				<form string="End of contract" version="7.0">
+                    <!-- End sponsorship form -->
+                    <group>
+                        <group>
+                            <field name="contract_id" readonly="1"/>
+                            <field name="child_id" readonly="1"/>
+                            <field name="end_date"/>
+                            <field name="end_reason"/>
+                        </group>
+                        <group/>
+
+                        <group>
+                            <!-- Child Exit Details form -->
+                            <field name="gp_exit_reason" attrs="{'required': [('end_reason', '=', '1')],
+                                                                 'invisible': [('end_reason', '!=', '1')]}"/>
+                        </group>
+                        <group/>
+
+                        <group attrs="{'invisible': [('end_reason', '!=', '4')]}">
+                            <!-- Transfer child form -->
+                            <field name="do_transfer" />
+                            <field name="transfer_country_id" options="{'limit': 20}"
+                                   attrs="{'invisible': [('do_transfer', '=', False)],
+                                           'required': [('do_transfer', '=', True)]}" />
+                        </group>
+                        <group/>
+                    </group>
+
+                    <!-- Action buttons in footer -->
+					<footer>
+						<button name="end_contract" string="End contract" 
+                                type="object" class="oe_highlight"/>
+                        <button string="Cancel" class="oe_link" special="cancel"/>
+					</footer>
+				</form>
+			</field>
+		</record>
+
+        <record id="action_end_contract" model="ir.actions.act_window">
+             <field name="name">End contract</field>
+             <field name="res_model">end.contract.wizard</field>
+             <field name="src_model">recurring.contract</field>
+             <field name="type">ir.actions.act_window</field>
+             <field name="view_type">form</field>
+             <field name="view_mode">form</field>
+             <field name="view_id" ref="end_contract_view"/>
+             <field name="context">{'active_id': active_id}</field>
+             <field name="target">new</field>
+       </record>
+    </data>
+</openerp>
