+++ contract_compassion/wizard/__init__.py
@@ -12,3 +12,4 @@
 from . import recurring_invoicer_wizard
 from . import contract_activation_wizard
 from . import install_contract_compassion
+from . import end_contract_wizard
